requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fonttools==4.49.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
importlib-resources==6.1.1
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.42.0
matplotlib==3.8.3
numba==0.59.0
numpy==1.26.4
openai==1.12.0
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pydantic==2.6.1
pydantic_core==2.16.2
pynndescent==0.5.11
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
threadpoolctl==3.3.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
umap-learn==0.5.5
zipp==3.17.0
tabulate==0.9.0
